moved GameState into utility.rs
