+++ config/communities.json
@@ -1,18 +1,7 @@
 {
     "0": {
         "id": 0,
-        "size": 1000000
-    },
-    "1": {
-        "id": 1,
-        "size": 1000000
-    },
-    "2": {
-        "id": 2,
-        "size": 1000000
-    },
-    "3": {
-        "id": 3,
-        "size": 1000000
+        "size": 100,
+        "money": 10000
     }
 }
