+++ src/utility.rs
@@ -1,11 +1,62 @@
 use crate::geography::*;
-use crate::population::*;
+use crate::population::Community;
 use serde::de;
+use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use std::error::Error;
 use std::fs::File;
 use std::io::BufReader;
 use std::path::Path;
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct GameState {
+    pub worlds: HashMap<usize, World>,
+    pub continents: HashMap<usize, Continent>,
+    pub regions: HashMap<usize, Region>,
+    pub provinces: HashMap<usize, Province>,
+    pub communities: HashMap<usize, Community>,
+}
+
+impl GameState {
+    pub fn new() -> GameState {
+        GameState {
+            worlds: HashMap::new(),
+            continents: HashMap::new(),
+            regions: HashMap::new(),
+            provinces: HashMap::new(),
+            communities: HashMap::new(),
+        }
+    }
+
+    pub fn init_from_config() -> GameState {
+        let mut game_state = GameState::new();
+        game_state.worlds = Config::load_config("config/worlds.json").unwrap();
+        game_state.continents = Config::load_config("config/continents.json").unwrap();
+        game_state.regions = Config::load_config("config/regions.json").unwrap();
+        game_state.provinces = Config::load_config("config/provinces.json").unwrap();
+        game_state.communities = Config::load_config("config/communities.json").unwrap();
+        game_state
+    }
+
+    pub fn add_world(self: &mut Self, id: usize, name: String) {
+        self.worlds.insert(id, World::new(id, name));
+    }
+
+    pub fn add_continent(self: &mut Self, id: usize, name: String) {
+        self.continents.insert(id, Continent::new(id, name));
+    }
+
+    pub fn add_region(self: &mut Self, id: usize, name: String) {
+        self.regions.insert(id, Region::new(id, name));
+    }
+    pub fn add_province(self: &mut Self, id: usize, name: String, resource_id: usize) {
+        self.provinces
+            .insert(id, Province::new(id, name, resource_id));
+    }
+    pub fn add_community(self: &mut Self, id: usize, size: usize, money: usize) {
+        self.communities.insert(id, Community::new(id, size, money));
+    }
+}
 
 pub struct Config;
